requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiobotocore==2.18.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
anndata==0.11.3
annotated-types==0.7.0
array_api_compat==1.10.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
biopython==1.85
botocore==1.36.1
cellxgene-census==1.15.0
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses==0.6
datasets==2.19.2
debugpy==1.8.12
decorator==5.1.1
deepchem==2.5.0
dill==0.3.8
et_xmlfile==2.0.0
evaluate==0.4.2
executing==2.1.0
fastprogress==1.0.3
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.3.1
fuzzywuzzy==0.18.0
gast==0.6.0
gensim==4.3.3
gget==0.29.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
huggingface-hub==0.27.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
legacy-api-wrap==1.4.1
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mysql-connector-python==9.1.0
namex==0.0.8
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyarrow-hotfix==0.6
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
PyTDC==1.1.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rdkit==2023.9.6
regex==2024.11.6
requests==2.32.3
rich==13.9.4
s3fs==2024.3.1
safetensors==0.5.2
scanpy==1.10.4
scikit-learn==1.2.2
scipy==1.13.1
seaborn==0.13.2
session_info==1.0.0
six==1.17.0
smart-open==7.1.0
SmilesPE==0.0.3
somacore==1.0.11
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiledb==0.29.1
tiledbsoma==1.11.4
tokenizers==0.19.1
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.43.4
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
